New optimized fuel and ignition schedules.
Some explaining remarks:
1.Separate dwell limiters is no longer needed because of rooted out problems with schedulers. Now schedulers itself act as dwell limiters.
2. Replace fixed cranking ignition functions also with new schedulers (in the decoders.ino).

+++ speeduino/globals.ino
@@ -151,7 +151,7 @@ volatile uint16_t ignitionCount; /**< The count of ignition events that have tak
   byte tertiaryTriggerEdge;
 #endif
 int CRANK_ANGLE_MAX = 720;
-int CRANK_ANGLE_MAX_IGN = 360;
+int16_t CRANK_ANGLE_MAX_IGN = 360;
 int CRANK_ANGLE_MAX_INJ = 360; ///< The number of crank degrees that the system track over. 360 for wasted / timed batch and 720 for sequential
 volatile uint32_t runSecsX10;
 volatile uint32_t seclx10;
